chore: more some types to avoid shat /src depends on /scripts
